TWINE: moved resource management files into own subdir
