+++ engines/twine/holomap.cpp
@@ -25,11 +25,11 @@
 #include "common/types.h"
 #include "twine/audio/sound.h"
 #include "twine/scene/gamestate.h"
-#include "twine/hqr.h"
+#include "twine/resources/hqr.h"
 #include "twine/menu/interface.h"
 #include "twine/renderer/renderer.h"
 #include "twine/renderer/screens.h"
-#include "twine/resources.h"
+#include "twine/resources/resources.h"
 #include "twine/scene/scene.h"
 #include "twine/text.h"
 #include "twine/twine.h"
