+++ engines/twine/renderer/redraw.cpp
@@ -31,14 +31,14 @@
 #include "twine/scene/collision.h"
 #include "twine/debugger/debug_scene.h"
 #include "twine/scene/grid.h"
-#include "twine/hqr.h"
+#include "twine/resources/hqr.h"
 #include "twine/input.h"
 #include "twine/menu/interface.h"
 #include "twine/menu/menu.h"
 #include "twine/scene/movements.h"
 #include "twine/renderer/renderer.h"
 #include "twine/renderer/screens.h"
-#include "twine/resources.h"
+#include "twine/resources/resources.h"
 #include "twine/scene/scene.h"
 #include "twine/text.h"
