+++ engines/twine/resources/hqr.h
