+++ engines/twine/resources/lzss.h
