+++ engines/twine/scene/actor.cpp
@@ -25,11 +25,11 @@
 #include "common/system.h"
 #include "common/textconsole.h"
 #include "twine/audio/sound.h"
-#include "twine/hqr.h"
+#include "twine/resources/hqr.h"
 #include "twine/parser/entity.h"
 #include "twine/renderer/renderer.h"
 #include "twine/renderer/screens.h"
-#include "twine/resources.h"
+#include "twine/resources/resources.h"
 #include "twine/scene/animations.h"
 #include "twine/scene/extra.h"
 #include "twine/scene/gamestate.h"
