+++ engines/twine/scene/extra.cpp
@@ -28,7 +28,7 @@
 #include "twine/menu/interface.h"
 #include "twine/renderer/redraw.h"
 #include "twine/renderer/renderer.h"
-#include "twine/resources.h"
+#include "twine/resources/resources.h"
 #include "twine/scene/actor.h"
 #include "twine/scene/collision.h"
 #include "twine/scene/gamestate.h"
