+++ engines/twine/text.cpp
@@ -28,13 +28,13 @@
 #include "common/str.h"
 #include "common/system.h"
 #include "twine/audio/sound.h"
-#include "twine/hqr.h"
+#include "twine/resources/hqr.h"
 #include "twine/input.h"
 #include "twine/menu/interface.h"
 #include "twine/menu/menu.h"
 #include "twine/renderer/renderer.h"
 #include "twine/renderer/screens.h"
-#include "twine/resources.h"
+#include "twine/resources/resources.h"
 #include "twine/scene/scene.h"
 #include "twine/twine.h"
